requirements.txt
Django==3.1.7
django-timezone-field==4.1.1
django-celery-beat==2.2.0
django-tenants==3.2.1
tenant-schemas-celery==1.0.1
